Analysis-Python/house.gov-representatives/requirements.txt
streamlit
bs4
seaborn
plotly
scipy
kaleido
vega_datasets
requests
pandas
matplotlib.pyplot
altair